rearranged files and got file finder working
