+++ src/lib/lib/extract-patterns-from-file.mjs
