+++ src/lib/lib/process-file-patterns.mjs
