+++ src/lib/lib/process-gitignore.mjs
