+++ src/lib/lib/process-package-ignores.mjs
