+++ src/lib/lib/test/data/test-gitignore-negative.txt
