+++ src/lib/lib/test/extract-patterns-from-file.test.mjs
