+++ src/lib/lib/test/mock-read-file.mjs
