+++ src/lib/lib/test/process-file-patterns.test.mjs
