+++ src/lib/lib/test/process-package-ignores.test.mjs
